retirement_planner/requirements.txt
# Core dependencies
plotext>=5.2.0  # Pure Python plotting library
PyYAML>=6.0  # For YAML config support
